internet gateway: clean up by tag, not by linked VPC
An IGW can exist unlinked from a VPC, and as a part of our cleanup we
unlink the IGW from its VPC. If we fail partway through the cleanup or
the IGW was never linked to a VPC we would miss it.

Now we search for IGWs by their tag. Maybe we should do both - search by
tag and by linked VPC, but the premise of this utility is that your
resources are properly tagged.